async fuction refactoring
