+++ main.py
@@ -4,6 +4,7 @@ import logging
 import traceback
 import uuid
 
+# from async_generator import asynccontextmanager
 from fastapi import FastAPI, File, UploadFile, HTTPException
 from fastapi.responses import FileResponse
 import uvicorn
@@ -11,54 +12,74 @@ from PIL import Image
 
 from model import VTryOnModel
 from config.config import opt
-import preprocess
+import preprocessor
 
-
-# config = Configuration()
 logging.basicConfig(filename='error.log', level=logging.DEBUG)
 
+# ml_models = {}
+
+# @asynccontextmanager
+# async def lifespan(app: FastAPI):
+#     # Load the model
+#     ml_models["vton"] = VTryOnModel(opt.device)
+#     yield
+    
+#     # Clean up the model and release the resources
+#     ml_models.clear()
+
+# app = FastAPI(lifespan=lifespan)
 
 app = FastAPI()
+vton = VTryOnModel(opt.device)
 
-@app.post("/upload/")
-async def upload_image(cloth: UploadFile = File(...), person: UploadFile = File(...)):
+
+async def preprocessing(cloth_path, edge_path, person_path):
+    for path in [cloth_path, edge_path, person_path]:
+        im = Image.open(path)
+        if im.size != (192, 256):
+            preprocessor.resize(path)
+
+    c_tensor, e_tensor, p_tensor = preprocessor.img_to_tensor(cloth_path, edge_path, person_path)
+
+    return c_tensor, e_tensor, p_tensor
+
+
+async def inference(c_tensor, e_tensor, p_tensor):
+    vton.infer(c_tensor, e_tensor, p_tensor)
+    result_image_path = "results/output.jpg"
+    
+    return FileResponse(result_image_path, media_type="image/jpeg")
+
+
+@app.post("/upload-images")
+async def upload_person(cloth: UploadFile = File(...), person: UploadFile = File(...) ):
     try:
-        # Model Initialization
-        vton = VTryOnModel(opt.device)
-
-        # Uploading Images
         cloth_path = opt.cloth_path + f"{uuid.uuid4()}.jpg"
-        with open(cloth_path, "wb") as img_file:
-            img_file.write(await cloth.read())
+        person_path = opt.person_path + f"{uuid.uuid4()}.jpg"
+        
+        with open(cloth_path, "wb+") as img_file:
+            img_file.write(cloth.file.read())
+        with open(person_path, "wb+") as img_file:
+            img_file.write(person.file.read())
 
         if not opt.edge_exist:
             edge_name = f"{uuid.uuid4()}.jpg"
             edge_path = opt.edge_path + edge_name
-            preprocess.generate_edge(edge_exist=False, img_path=cloth_path, output_name=edge_name)
+            preprocessor.generate_edge(edge_exist=False, img_path=cloth_path, output_name=edge_name)
         else:
             edge_path = "dataset/service_edge/ada6385c-a2e7-44a1-a48d-6b0b64d47963.jpg"
 
-        person_path = opt.person_path + f"{uuid.uuid4()}.jpg"
-        with open(person_path, "wb") as img_file:
-            img_file.write(await person.read())
-
-        # Preprocessing
-        for img_path in [cloth_path, edge_path, person_path]:
-            im = Image.open(img_path)
-            if im.size != (192, 256):
-                preprocess.resize(img_path)
-
-        c_tensor, e_tensor, p_tensor = preprocess.img_to_tensor(cloth_path, edge_path, person_path)
+        c_tensor, e_tensor, p_tensor = await preprocessing(cloth_path, edge_path, person_path)
         
-        # Inference
-        vton.infer(c_tensor, e_tensor, p_tensor)
+        await inference(c_tensor, e_tensor, p_tensor)
+
         result_image_path = "results/output.jpg"
-        
-        return FileResponse(result_image_path, media_type="image/jpeg")
+        return FileResponse(result_image_path, media_type="image/jpeg") 
 
     except Exception as e:
         logging.error(traceback.format_exc())
         raise HTTPException(status_code=500, detail=str(e))
+
 
 if __name__ == "__main__":
     uvicorn.run(app, host = "127.0.0.1", port=8000)
